Configuration syncing, compile all fix, compile on save
